Add support for running MCP server initialization code
Clients will typically invoke ListTools right after client initialization is complete. We don't attempt to run this code as a notification handler because those don't have access to the server instance and therefore are not very useful.

Running just before tools list is a good spot for post-initialized and before things can actually start running.

Note that doing elicitation in this context won't work, but other notifications/logging to the client should be perfectly functional at this stage.